gwt 2.3 m1 ; deleted server.datasource and friends
git-svn-id: http://192.168.0.243:4435/svn/bee/trunk@204 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/shared/utils/Grego.java
@@ -27,7 +27,7 @@ public class Grego {
 
   private static final int THURSDAY = 5;
 
-  public static int dayOfWeek(long day) {
+  public static int dayOfWeek(int day) {
     long[] remainder = new long[1];
     floorDivide(day + THURSDAY, 7, remainder);
